@@ -3,21 +3,21 @@
     "version": "0.1.0",
     "private": true,
     "devDependencies": {
-        "react-scripts": "0.8.5"
+        "react-scripts": "3.1.0"
     },
     "dependencies": {
-        "firebase": "3.6.8",
-        "history": "4.5.0",
-        "react": "15.4.2",
-        "react-cookie": "1.0.4",
-        "react-dom": "15.4.2",
-        "react-google-charts": "1.2.1",
-        "react-redux": "5.0.2",
-        "react-router": "3.0.0",
-        "react-router-redux": "4.0.7",
-        "redux": "3.6.0",
-        "redux-logger": "2.7.0",
-        "redux-thunk": "2.1.0"
+        "firebase": "6.3.5",
+        "history": "4.9.0",
+        "react": "16.9.0",
+        "react-cookie": "4.0.1",
+        "react-dom": "16.9.0",
+        "react-google-charts": "3.0.14",
+        "react-redux": "7.1.0",
+        "react-router": "3.2.3",
+        "react-router-redux": "4.0.8",
+        "redux": "4.0.4",
+        "redux-logger": "3.0.6",
+        "redux-thunk": "2.3.0"
     },
     "scripts": {
         "start": "react-scripts start",
