@@ -1,25 +1,39 @@
-import React, {Component } from "react"
-import {connect} from 'react-redux'
-import {RegForm, AnsButtons} from '../components'
-import {Constants} from '../utils'
-import cookie from 'react-cookie'
+import React, {Component } from "react";
+import {connect} from 'react-redux';
+import {RegForm, AnsButtons} from '../components';
+import {Constants} from '../utils';
+import { instanceOf } from 'prop-types';
+import { withCookies, Cookies } from 'react-cookie';
 
 // 回答ページ用のコンテナ、クイズ参加者はスマートフォンで操作
 class SmartDevice extends Component {
+    static propTypes = {
+        cookies: instanceOf(Cookies).isRequired
+    };
+
+    constructor(props) {
+        super(props);
+
+        const { cookies } = props;
+        this.state = {
+            name: cookies.get(Constants.cookie.NAME)
+        };
+    }
 
     onRegister(name) {
+        const { cookies } = this.props;
         // クイズ参加者の名前をクッキーに保存
-        cookie.save(Constants.cookie.NAME, name, {
+        cookies.set(Constants.cookie.NAME, name, {
             path: '/',
             maxAge: Constants.cookie.MAX_AGE
-        })
-        this.forceUpdate()
+        });
+        this.forceUpdate();
     }
 
     render() {
-        const {globalStatus, quizzes} = this.props
-        const currentQuiz = quizzes[globalStatus.currentQuizIndex]
-        const name = cookie.load(Constants.cookie.NAME)
+        const {globalStatus, quizzes, cookies} = this.props;
+        const currentQuiz = quizzes[globalStatus.currentQuizIndex];
+        const name = cookies.get(Constants.cookie.NAME);
 
         return (
             <div className="smartDeviceFrame">
@@ -32,7 +46,7 @@ class SmartDevice extends Component {
                         globalStatus={globalStatus}
                         currentQuiz={currentQuiz}/>}
             </div>
-        )
+        );
     }
 }
 
@@ -43,4 +57,4 @@ const mapStateToProps = (state) => {
     }
 }
 
-export default connect(mapStateToProps)(SmartDevice)
+export default connect(mapStateToProps)(withCookies(SmartDevice));
