@@ -1,11 +1,4 @@
-import App from './App'
-import SmartDevice from './SmartDevice'
-import Controller from './Controller'
-import Screen from './Screen'
-
-module.exports = {
-    App,
-    SmartDevice,
-    Controller,
-    Screen
-}
+export { default as App }  from './App';
+export { default as SmartDevice } from './SmartDevice';
+export { default as Controller } from './Controller';
+export { default as Screen } from './Screen';
